Add -drain-poll-interval, fix tests
