+++ config/config.go
@@ -25,6 +25,7 @@ const (
 	defaultDrainMinHealthySiblingLifetime   = "1h"
 	defaultDrainMinUnhealthySiblingLifetime = "6h"
 	defaultDrainForceEvictInterval          = "5m"
+	defaultDrainPollInterval                = "30s"
 	defaultUpdateStrategy                   = "rolling"
 )
 
@@ -108,6 +109,7 @@ func (cfg *LifecycleManagerConfig) ParseFlags() string {
 	kingpin.Flag("drain-min-healthy-sibling-lifetime", "Minimum lifetime of healthy pods in the same PDB as the one considered for forced termination.").Default(defaultDrainMinHealthySiblingLifetime).DurationVar(&cfg.UpdateStrategy.MinHealthyPDBSiblingLifetime)
 	kingpin.Flag("drain-min-unhealthy-sibling-lifetime", "Minimum lifetime of unhealthy pods in the same PDB as the one considered for forced termination.").Default(defaultDrainMinUnhealthySiblingLifetime).DurationVar(&cfg.UpdateStrategy.MinUnhealthyPDBSiblingLifetime)
 	kingpin.Flag("drain-force-evict-interval", "Interval between forced terminations of pods on the same node.").Default(defaultDrainForceEvictInterval).DurationVar(&cfg.UpdateStrategy.ForceEvictionInterval)
+	kingpin.Flag("drain-poll-interval", "Interval between drain attempts.").Default(defaultDrainPollInterval).DurationVar(&cfg.UpdateStrategy.PollInterval)
 	kingpin.Flag("update-strategy", "Update strategy to use when updating node pools.").Default(defaultUpdateStrategy).EnumVar(&cfg.UpdateStrategy.Strategy, "rolling")
 	kingpin.Flag("remove-volumes", "Remove EBS volumes when decommissioning.").BoolVar(&cfg.RemoveVolumes)
 	kingpin.Flag("environment-order", "Roll out channel updates to the environments in a specific order.").StringsVar(&cfg.EnvironmentOrder)
